requirements.txt
asgiref==3.8.1
crontab==1.0.1
Django==4.2.11
django-crontab==0.7.1
djangorestframework==3.15.1
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1